Change the api request to get all mod data at once
Addresses the issue of getting hit with code 429 by
requesting every mod individually. This had also caused
some irregularities in the database based on which
queued requests were rejected. Users deleting the database
is advisable.

A new issue with this change is that the mod list tab does
not update mods as deprecated/etc until you go back to the
home page and reload the file. There is also visible
freezing when switching to the mod tab. Both of these may
have to do with FutureBuilder and how/when the data is
computed, but I have not gotten to the bottom of this.