Convert the ModManager to a provider
The data of this class can be modified by multiple screens
and relying on setState is not sufficient anymore for
update. For example, from the mod list tab going into the
settings to change, e.g., the problematic list, and then
back would reflect any changes until someone where to move
away from the tab and back to it again.

The FutureBuilder is not necessary anymore, which also
seemed to cause some lag every time when moving to the mod
tab. However, I have noticed that mods not currently in
the database won't trigger a UI when the logger async
task. I think it's because the mod tab has not yet
subscribed to the ModManager to be notified for changes.