Fix deprecated mods not properly displaying their status
When accessing a mod's status from Thunderstore, the
deprecation check was incorrect.

Furthermore, during my tests I was working with some logs
that apparently loaded a mod more than once. A mod's name
reference is overridden with its latest instance, so the
only one that would show as deprecated once properly
loaded would be the last one, which might be off-screen
and give the impression it was not properly flagged. In
order to avoid such a confusion, a mod's first encounter
is mapped to its name even though such "double mods"
should be quite rare if non-existent at all.