Add the game directory to the summary is suspected cracked
Steam should always have the "steamapps/common/" pattern,
while Epic Games does allow to install anywhere, resulting
in a possible false positive. The inclusion of this line
should be more about drawing attention to the possibility
and to further investigate it, than outright flag it
accordingly.

Also converted a bunch of var to final.