Fix when a mod name appears multiple times not updating all copies
This is a rare issue that occurs for mods that contain multiple plugins. The
plugins are loaded individually, but are all preceded by the same
'TS Manifest: <mod name>' message which ends up registering the same mod multiple
times. When retrieving a mod by its name, we must not return just the first copy
encountered.