Make the app functional for the web
The commit is mostly concerned with loading a file and having a database version
that works. Both of these issues are solved with new libraries since isolates do
not work for the web, for example.

This is only to allow the app to make it to the console screen at least for
development. Any release, deployment, or other web-related issues will be dealt
with later.